Improved polymorphic implementation to handle virtual bases in a slightly UB way, given the loose specification of where virtual bases are stored. In reality I'm pretty sure that the offset to a virtual base is the same for the _same_ ultimate subclass.
TODO: Add tests for second-base and virtual base case to see that the offset calculation and usage works.

Also includes a stab at a cloning_ptr which is more pointer-like and uses a unique_ptr to store the data and a cloner to clone it. Requires virtual dtor which polymorphic doesn't.

For some reason I can't figure out cloning_ptr does not work in the forward-declared scenario that polymorphic works in, as displayed by the compilation test. That's why the cloning_ptr_test.cc program does not compile.