Debugged cloning_ptr.
It still does not support virtual bases though, which this PR's polymorphic should.

Next up is fixing this, and adding test cases to prove it works.

Next step is then to use a requires clause or similar to get rid of the 'offset' overhead when T is not a virtual base of U. Note that this offset is known at compile time for all U, T pairs but there is no way to get at it in C++. Well maybe, using constexpr funtcion that does the offset calculation, but that would trigger an error as reinterpret_cast is not available at compile time.