Made a new cloning_ptr implementation that works with virtual bases, without doing strange offset arithmetic. Relies on a fat pointer though, but uses regular virtual methods.
This implementation could be made non-fat in the same way as polymorphic, by adding a T* pointer in impl_base instead of in cloning_ptr itself. The performance hit at each dereference is quite high.

Going back to offset arithmetic and storing the offset to add to the impl pointer to get to the T in its m_value member would allow (I think) assigning between cloning_ptr of different T without nesed control blocks. By inverting the offset to be to impl _from_ the T subobject the cost at indirection can simultaneously be 0.