Various fixes related to allocators
Fix several cases of `allocator_traits<Alloc>` where `Alloc` is not defined.

Remove `rebind_traits::` which is not valid, because it's missing a template argument list. It's unnecessary for calling `construct` and `destroy` anyway.

Remove notes giving normative requirements about T being a complete type, replace with _Mandates_ elements. I'm unsure about the `polymorphic` constructor that's intended to use the small buffer optimization. The note says completeness of T depends on allocator behaviour, but that's not true if it also depends on whether a small buffer optim is used (it depends on whether T fits in the buffer, which requires T to be complete!)

Fix `noexcept(see below)` in `indirect` synopsis which is just an unconditional `noexcept` on the detailed description (and the effects don't describe anything that can throw).