Mo fixes (#413)
* polymorphic move constructor should be noexcept(true)

The class synopsis doesn't match the detailed description.

* Fix self-assignment checks

The reference implementation compares addresses, not values, so it seems
that is the intention. Comparing values does nto even compile for
`polymoprhic` as it has not `operator==`.

* Use propagate_on_container_move_assignment in move assignment

* Add missing "not" to polymorphic::swap effects

* Remove unnecessary std:: qualification

* Add missing constructible constraints to polymorphic ctors

It's not enough to check that remove_cvref_t<U> is copy constructible,
you also need to check that it can actually be constructed from the
constructor's U argument. That type could be U& or U&& or const U&&
which are not checked by the is_copy_constructible trait.