Add implementation of polymorphic that uses no vtable (#571)
This is an amendment of https://github.com/jbcoe/value_types/pull/567
that introduces a parallel implementation of polymorphic rather than
replacing the existing one.

Co-authored with @toughengineer

---

Following discussion in #559 I replaced inheritance and virtual
functions with a single pointer to function.

With the current benchmarks I could not reliably measure the difference
in performance so I assumed that it is not significantly different.

I could however very easily measure the difference in resulting binary
size.
Here's the size of **polymorphic_test.exe** in KB:

|compiler|current size|size after change|difference|
|-|-|-|-|
|MSVC|921|904|-17|
|Clang|996|982|-14|

I did not investigate what exactly contributed to overhead, I bet it's
RTTI, and maybe vtables and individual functions a little bit.

Solution with a pointer to a single function (apart from binary size
compared to inheritance and virtual functions) is very ABI friendly, and
can be made extendable in an ABI compatible way.